updated tests and refactored into a test helper
