+++ invoices/get_test.go
@@ -3,8 +3,8 @@ package invoices
 import (
 	"testing"
 
-	dinero "github.com/eikc/dinero-go"
-	"github.com/eikc/dinero-go/dinerotest"
+	"github.com/eikc/dinero-go/internal"
+	"github.com/eikc/dinero-go/internal/dinerotest"
 )
 
 func TestGetInvoice_integration(t *testing.T) {
@@ -12,10 +12,7 @@ func TestGetInvoice_integration(t *testing.T) {
 		t.Skip(dinerotest.IntegrationTestText)
 	}
 
-	key, secret, apiKey, organizationID := dinerotest.GetClientKeysForIntegrationTesting()
-
-	c := dinero.NewClient(key, secret)
-	c.Authorize(apiKey, organizationID)
+	c := internal.GetClient()
 
 	if _, err := Get(c, "f25ed5df-3fb4-49b5-98ce-31711238fc10"); err != nil {
 		t.Error("We could not get an invoice, error: ", err)
